Add new module for preprocessing data
We have more than just plotting now, so better to refactor into a
separate module
